Make AnimationStateUpdate jobs run in parallel
